GH-219: removed bogus scope test
Due to considering instance members of function for constructor type,
some scope tests had to be adjusted.

In StaticScopingTest, the expectations were wrong, as the set up wrongly
uses an instance access to compute a static scope (cf. line 66 in
previous). Removed this test rather instead of fixing it, since there
are Xpect tests doing the same thing in a much more readable manner.