GH-219: refactored member scoping and union type scoping in particular
Several bugs revealed while working on GH-219. Refactored
MemberScopingHelper to make it clearer when a scope is "decorated", that
is, when filters are added to the original scope. Renamed
StaticAccessAwareMemberScope to StaticWriteAccessFilterScope and
improved documentation there to make purpose more clear. Also renamed
ComposedMemberScope to UnionMemberScope, since this is all that this
scope is doing at the moment and a scope for intersection will likely
look much different.