GH-13: Aligned existing tests to the new warnings.
In about 30 tests, union and intersection types where defined that do provoke one or more of the newly introduced warnings. In most of the cases, the union or intersection types consist of classes e.g. intersection{A,B} where A and B are classes. In these cases, the warning could be eliminated by turning the classes A and B into interfaces. Only in some cases, the newly introduced warnings have been added to the tests expectancy.