GH-212 moves getter/setter shadow-delegation to ApiImplStubGeneration
In general a validation enforces to provide getter/setter in pairs in
case of overriding only one in the presence of the other. But for
API-Stubs this does not hold and the purpose of the stub is to provide
an error-message. Hence we delegate the existing inherited setter or
getter to it's defining classifier.