IDE-2468: fix in handling of refs to const fields in compile-time expr
This commit makes sure that in a compile-time expression a reference to
a const field with a computed name raises an error even if the const
field is located in another file an the TField was obtained via the
index. This is required for consistency with the cases that the const
field is located in the same file or is located in another file but that
file does not have type information in the index.